refactored `Doc::TokensList` component to expose `Doc::TokenPreview` as standalone reusable component
