+++ website/app/components/doc/token-preview/index.hbs
@@ -1,0 +1,22 @@
+{{!
+  Copyright (c) HashiCorp, Inc.
+  SPDX-License-Identifier: MPL-2.0
+}}
+
+{{! template-lint-disable no-inline-styles }}
+{{#if this.colorPreviewStyle}}
+  <div class="doc-token-preview doc-token-preview--color" style={{this.colorPreviewStyle}}></div>
+{{else if this.backgroundImagePreviewStyle}}
+  <div class="doc-token-preview doc-token-preview--image" style={{this.backgroundImagePreviewStyle}}></div>
+{{else if this.fontPreviewStyle}}
+  <div class="doc-token-preview doc-token-preview--font" style={{this.fontPreviewStyle}}>Aa</div>
+{{else if this.sizePreviewStyle}}
+  <div class="doc-token-preview doc-token-preview--size" style={{this.sizePreviewStyle}}><span
+      class="doc-token-preview__value"
+    >{{this.token.value}}</span></div>
+{{else if this.boxShadowPreviewStyle}}
+  <div class="doc-token-preview doc-token-preview--boxshadow" style={{this.boxShadowPreviewStyle}}></div>
+{{else}}
+  <div class="doc-token-preview doc-token-preview--unknown"></div>
+{{/if}}
+{{! template-lint-enable no-inline-styles }}
