+++ website/app/components/doc/tokens-list/item.js
@@ -6,9 +6,8 @@
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
 import { tracked } from '@glimmer/tracking';
-import { htmlSafe } from '@ember/template';
 
-export default class DocTokenCardIndexComponent extends Component {
+export default class DocTokensListComponent extends Component {
   @tracked isExpanded = false;
 
   get token() {
@@ -34,58 +33,6 @@ export default class DocTokenCardIndexComponent extends Component {
 
   get isDeprecated() {
     return this.token.deprecated;
-  }
-
-  get colorPreviewStyle() {
-    const isColor =
-      this.token.value.startsWith('#') || this.token.value.startsWith('rgb');
-    return isColor
-      ? htmlSafe(`background-color: ${this.token.value}`)
-      : undefined;
-  }
-
-  get backgroundImagePreviewStyle() {
-    const isBackgroundImage = this.token.value.match(/url\("data:image\//);
-    let backgroundColor;
-    if (this.token.value.match(/fill='%23f{3,6}'/i)) {
-      backgroundColor = 'rgb(0 0 0 / 15%)';
-    } else {
-      backgroundColor = 'transparent';
-    }
-    return isBackgroundImage
-      ? htmlSafe(
-          `background-image: ${this.token.value}; background-color: ${backgroundColor}`
-        )
-      : undefined;
-  }
-
-  get fontPreviewStyle() {
-    if (this.token.type === 'font-size') {
-      return htmlSafe(`font-size: ${this.token.value}`);
-    } else if (
-      this.token.name.startsWith('token-typography') &&
-      (this.token.name.includes('font-stack') ||
-        this.token.name.includes('font-family'))
-    ) {
-      return htmlSafe(`font-family: ${this.token.value}`);
-    } else if (this.token.name.startsWith('token-typography-font-weight')) {
-      return htmlSafe(`font-weight: ${this.token.value}`);
-    } else {
-      return undefined;
-    }
-  }
-
-  get sizePreviewStyle() {
-    const isSize =
-      this.token.type === 'size' && this.token.value.endsWith('px');
-    return isSize ? htmlSafe(`height: ${this.token.value}`) : undefined;
-  }
-
-  get boxShadowPreviewStyle() {
-    const isBoxShadow = this.token.name.endsWith('box-shadow');
-    return isBoxShadow
-      ? htmlSafe(`box-shadow: ${this.token.value}`)
-      : undefined;
   }
 
   @action
