+++ website/app/styles/doc-components/index.scss
@@ -20,5 +20,6 @@
 @import "meta-row";
 @import "placeholder";
 @import "table-of-contents";
+@import "token-preview";
 @import "tokens-list";
 @import "vars-list";
