+++ website/app/styles/doc-components/token-preview.scss
@@ -1,0 +1,103 @@
+/**
+ * Copyright (c) HashiCorp, Inc.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+// TOKEN PREVIEW
+
+@use "../typography/mixins" as *;
+
+.doc-token-preview {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 24px;
+  height: 24px;
+  border-radius: 2px;
+}
+
+.doc-token-preview--color {
+  box-shadow: inset 0 0 1px 0 rgb(0, 0, 0, 25%);
+}
+
+.doc-token-preview--image {
+  background-repeat: no-repeat;
+  background-position: center center;
+  box-shadow: inset 0 0 1px 0 rgb(0, 0, 0, 25%);
+
+  &::before {
+    position: absolute;
+    z-index: -1;
+    overflow: hidden;
+    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 2 2'%3E%3Cpath d='M1 2V0h1v1H0v1z' fill-opacity='.05'/%3E%3C/svg%3E");
+    background-size: cover;
+    // use this to fine tune the size of the checkered pattern
+    background-size: 8px 8px;
+    border-radius: inherit;
+    content: "";
+    inset: 0;
+  }
+}
+
+.doc-token-preview--font {
+  color: var(--doc-color-black);
+}
+
+.doc-token-preview--size {
+  --color: #ff9a8b;
+  --height: var(--token-value-height); // this is set via JavaScript
+
+  .doc-token-preview__value {
+    @include doc-font-family("mono");
+    position: absolute;
+    top: calc(50% - 4px);
+    left: 10px;
+    color: var(--color);
+    font-size: 8px;
+    line-height: 8px;
+  }
+
+  &::before {
+    position: absolute;
+    top: 50%;
+    left: 0;
+    width: 7px;
+    height: var(--height);
+    max-height: 24px;
+    border-top: 1px solid var(--color);
+    border-bottom: 1px solid var(--color);
+    transform: translateY(-50%);
+    content: "";
+  }
+
+  &::after {
+    position: absolute;
+    top: 50%;
+    left: 3px;
+    height: var(--height);
+    max-height: 24px;
+    border-left: 1px solid var(--color);
+    transform: translateY(-50%);
+    content: "";
+  }
+}
+
+.doc-token-preview--boxshadow {
+  --box-shadow: var(--token-value-box-shadow); // this is set via JavaScript
+  box-shadow: var(--box-shadow);
+}
+
+.doc-token-preview--unknown {
+  &::after {
+    position: absolute;
+    top: calc(50% - 16px);
+    left: 50%;
+    width: 1px;
+    height: 32px;
+    background-color: #999;
+    transform: rotate(45deg);
+    transform-origin: center center;
+    content: "";
+  }
+}
