+++ website/app/styles/doc-components/tokens-list.scss
@@ -5,8 +5,7 @@
 
 // TOKENS LIST
 
-@use "../../typography/mixins" as *;
-@use "../../breakpoints" as breakpoint;
+@use "../breakpoints" as breakpoint;
 
 
 .doc-tokens-list {
@@ -47,79 +46,6 @@
     justify-content: center;
     border-radius: 3px 0 0 3px;
   }
-}
-
-.doc-tokens-list__preview-color {
-  width: 24px;
-  height: 24px;
-  border-radius: 2px;
-  box-shadow: inset 0 0 1px 0 rgb(0, 0, 0, 25%);
-}
-
-.doc-tokens-list__preview-image {
-  position: relative;
-  width: 24px;
-  height: 24px;
-  background-repeat: no-repeat;
-  background-position: center center;
-  border-radius: 2px;
-  box-shadow: inset 0 0 1px 0 rgb(0, 0, 0, 25%);
-
-  &::before {
-    position: absolute;
-    z-index: -1;
-    overflow: hidden;
-    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 2 2'%3E%3Cpath d='M1 2V0h1v1H0v1z' fill-opacity='.05'/%3E%3C/svg%3E");
-    background-size: cover;
-    // use this to fine tune the size of the checkered pattern
-    background-size: 8px 8px;
-    border-radius: inherit;
-    content: "";
-    inset: 0;
-  }
-}
-
-.doc-tokens-list__preview-font {
-  color: var(--doc-color-black);
-}
-
-.doc-tokens-list__preview-size {
-  --border-color: #ff9a8b;
-  position: relative;
-  display: block;
-  width: 1px;
-  border-left: 1px solid var(--border-color);
-
-  &::before {
-    position: absolute;
-    top: 0;
-    right: -6px;
-    left: -6px;
-    border-top: 1px solid var(--border-color);
-    content: "";
-  }
-
-  &::after {
-    position: absolute;
-    right: -6px;
-    bottom: 0;
-    left: -6px;
-    border-bottom: 1px solid var(--border-color);
-    content: "";
-  }
-}
-
-.doc-tokens-list__preview-boxshadow {
-  width: 24px;
-  height: 24px;
-  border-radius: 2px;
-}
-
-.doc-tokens-list__preview-unknown {
-  position: relative;
-  width: 24px;
-  height: 24px;
-  background: linear-gradient(-45deg, #fff 0%, #fff 48%, #000 50%, #fff 48%, #fff 100%);
 }
 
 // CONTENT
